+++ src/util.ts
@@ -293,3 +293,22 @@ export function deepMerge(lhs: any, rhs: any): any {
 function pick(lhs: any, rhs: any): any {
   return lhs === undefined ? rhs : lhs;
 }
+
+export function getMods(api: types.IExtensionApi, modType: string): types.IMod[] {
+  const state = api.getState();
+  const mods = util.getSafe(state, ['persistent', 'mods', GAME_ID], {});
+  return Object.values(mods).filter((mod: types.IMod) => mod.type === modType) as types.IMod[];
+}
+
+export async function findModByFile(api: types.IExtensionApi, modType: string, fileName: string): Promise<types.IMod> {
+  const mods = getMods(api, modType);
+  const installationPath = selectors.installPathForGame(api.getState(), GAME_ID);
+  for (const mod of mods) {
+    const modPath = path.join(installationPath, mod.installationPath);
+    const files = await walkPath(modPath);
+    if (files.find(file => file.filePath.endsWith(fileName))) {
+      return mod;
+    }
+  }
+  return undefined;
+}