+++ src/views/Settings.tsx
@@ -1,8 +1,10 @@
 import * as React from 'react';
 import { ControlLabel, FormGroup, Panel, HelpBlock } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
-import { useSelector } from 'react-redux';
-import { Toggle } from 'vortex-api';
+import { useSelector, useStore } from 'react-redux';
+import { Toggle, util } from 'vortex-api';
+
+import { setManageLoadOrder } from '../actions/settings';
 
 import { NS, JUNCTION_TEXT } from '../common';
 
@@ -12,6 +14,7 @@ interface IBaseProps {
 
 interface IConnectedProps {
   enableDirectoryJunction: boolean;
+  manageLoadOrder: boolean;
 }
 
 type IProps = IBaseProps;
@@ -19,16 +22,31 @@ type IProps = IBaseProps;
 export default function Settings(props: IProps) {
   const { t } = useTranslation(NS);
   const { onSetDirectoryJunction } = props;
+  const store = useStore();
   const onToggle = React.useCallback((newVal) => {
     onSetDirectoryJunction(newVal);
-  }, [onSetDirectoryJunction])
-  const { enableDirectoryJunction } = useSelector(mapStateToProps);
+  }, [onSetDirectoryJunction]);
+
+  const onSetManageLO = React.useCallback((newVal) => {
+    store.dispatch(setManageLoadOrder(newVal));
+  }, [store]);
+
+  const { enableDirectoryJunction, manageLoadOrder } = useSelector(mapStateToProps);
   return (
     <form>
       <FormGroup controlId='default-enable'>
         <Panel>
           <Panel.Body>
             <ControlLabel>{t('Starfield')}</ControlLabel>
+            <Toggle
+              checked={manageLoadOrder}
+              onToggle={onSetManageLO}
+            >
+              {t('Manage Load Order')}
+            </Toggle>
+            <HelpBlock>
+              {t('This will tell Vortex to enable plugin load order management via the load order page.')}
+            </HelpBlock>
             <Toggle
               checked={enableDirectoryJunction}
               onToggle={onToggle}
@@ -47,6 +65,7 @@ export default function Settings(props: IProps) {
 
 function mapStateToProps(state: any): IConnectedProps {
   return {
-    enableDirectoryJunction: state.settings.starfield.enableDirectoryJunction,
+    enableDirectoryJunction: util.getSafe(state, ['settings', 'starfield', 'enableDirectoryJunction'], false),
+    manageLoadOrder: util.getSafe(state, ['settings', 'starfield', 'manageLoadOrder'], true),
   };
 }